+++ flight-data-recorder-impl/src/test/java/com/dwarfeng/fdr/impl/fuhconf/KafkaConfiguration.java
@@ -1,4 +1,4 @@
-package com.dwarfeng.fdr.impl.conf;
+package com.dwarfeng.fdr.impl.fuhconf;
 
 import com.dwarfeng.subgrade.sdk.kafka.serialize.FastJsonKafkaDeserializer;
 import org.apache.kafka.clients.consumer.ConsumerConfig;