+++ flight-data-recorder-impl/src/test/java/com/dwarfeng/fdr/impl/service/RealtimeValueMaintainServiceImplTest.java
@@ -60,5 +60,4 @@ public class RealtimeValueMaintainServiceImplTest {
             pointMaintainService.delete(parentPoint.getKey());
         }
     }
-
 }